Extract ImageBase class into separate module
